pyproject.toml
[tool.poetry]
name = "pixel-monitor"
version = "0.1.0"
description = ""
authors = ["Kristobal Junta <junta.kristobal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
parsel = "^1.5.2"
requests = "^2.23.0"
pyTelegramBotAPI = "^3.6.7"
python-dotenv = "^0.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
